Port mirrorhouse to SDL2 and restructure project
Build it again with the vendored SDL2-2.0.10. The display/input code in main.cpp was ported from the removed SDL 1.2 API: SDL_SetVideoMode -> SDL_CreateWindow + SDL_GetWindowSurface (windowed 1600x900), SDL_Flip -> SDL_UpdateWindowSurface, SDL_WM_GrabInput -> SDL_SetRelativeMouseMode. Guarded the M_PI define (now provided by SDL_stdinc.h) and silenced the Colour narrowing warnings.

Restructure: sources moved to src/, stub main.cpp and prebuilt SDL 1.2 binaries removed (preserved in history), project/filters updated. SDL.props now copies SDL2.dll to the output dir on build. Builds clean (0 warnings) for Debug|x64 and Release|x64; runs.

+++ src/colour.h
@@ -30,9 +30,9 @@ struct Colour
 	Colour operator*(double factor) const
 	{
 		return Colour(
-			std::min(std::max(red * factor, 0.0), 255.0),
-			std::min(std::max(blue * factor, 0.0), 255.0),
-			std::min(std::max(green * factor, 0.0), 255.0)
+			static_cast<uint8_t>(std::min(std::max(red * factor, 0.0), 255.0)),
+			static_cast<uint8_t>(std::min(std::max(blue * factor, 0.0), 255.0)),
+			static_cast<uint8_t>(std::min(std::max(green * factor, 0.0), 255.0))
 		);
 	}
